requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
SQLAlchemy==1.4.44
pydantic==1.10.2
alembic==1.7.7
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv
python-multipart